ref: move out metadata compliant fuzz tests
