+++ src/tests/resistance/fuzzing_stress.rs
@@ -17,10 +17,10 @@ use crate::{
         corrupt_message::{
             bad_checksum::encode_messages_and_corrupt_checksum,
             incorrect_length::encode_messages_and_corrupt_body_length_field,
-            slightly_corrupted_messages,
+            random_payload::metadata_compliant_random_bytes, slightly_corrupted_messages,
         },
         default_fuzz_messages,
-        fuzzing_random_bytes::{metadata_compliant_random_bytes, random_bytes},
+        fuzzing_random_bytes::random_bytes,
         seeded_rng,
         zeroes::zeroes,
         COMMANDS_WITH_PAYLOADS,
